music-recommender/requirements.txt
pandas
google-cloud-bigquery
google-cloud-storage
faiss-cpu
hdbscan
scikit-learn 
sqlalchemy
dotenv
psycopg2-binary
